LPS-56029 Move OSGi AutoLogin and AuthVerifier implementations into respective auto.login and auth.verifier packages
