+++ modules/portal-security/portal-security-auth/src/com/liferay/portal/security/auth/verifier/TunnelingServletAuthVerifier.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.auth.internal;
+package com.liferay.portal.security.auth.verifier;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
