+++ modules/portal-security/portal-security-sso-cas/src/com/liferay/portal/security/sso/cas/auto/login/CASAutoLogin.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.cas;
+package com.liferay.portal.security.sso.cas.auto.login;
 
 import com.liferay.portal.NoSuchUserException;
 import com.liferay.portal.kernel.exception.SystemException;
