+++ modules/portal-security/portal-security-sso-ntlm/src/com/liferay/portal/security/sso/ntlm/auto/login/NtlmAutoLogin.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm.security.auth;
+package com.liferay.portal.security.sso.ntlm.auto.login;
 
 import com.liferay.portal.kernel.security.auto.login.AutoLogin;
 import com.liferay.portal.kernel.security.auto.login.BaseAutoLogin;
