+++ modules/portal-security/portal-security-sso-openid/src/com/liferay/portal/security/sso/openid/auto/login/OpenIdAutoLogin.java
@@ -12,17 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.facebook.connect.auth;
+package com.liferay.portal.security.sso.openid.auto.login;
 
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.facebook.FacebookConnect;
+import com.liferay.portal.kernel.openid.OpenId;
 import com.liferay.portal.kernel.security.auto.login.AutoLogin;
 import com.liferay.portal.kernel.security.auto.login.BaseAutoLogin;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.model.User;
-import com.liferay.portal.security.sso.facebook.connect.constants.FacebookConnectWebKeys;
 import com.liferay.portal.service.UserLocalServiceUtil;
 import com.liferay.portal.util.PortalUtil;
 
@@ -34,10 +30,10 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Wilson Man
+ * @author Jorge Ferrer
  */
 @Component(immediate = true, service = AutoLogin.class)
-public class FacebookConnectAutoLogin extends BaseAutoLogin {
+public class OpenIdAutoLogin extends BaseAutoLogin {
 
 	@Override
 	protected String[] doLogin(
@@ -46,58 +42,36 @@ public class FacebookConnectAutoLogin extends BaseAutoLogin {
 
 		long companyId = PortalUtil.getCompanyId(request);
 
-		if (!_facebookConnect.isEnabled(companyId)) {
+		if (!_openId.isEnabled(companyId)) {
 			return null;
 		}
 
-		User user = getUser(request, companyId);
+		HttpSession session = request.getSession();
 
-		if (user == null) {
+		Long userId = (Long)session.getAttribute(WebKeys.OPEN_ID_LOGIN);
+
+		if (userId == null) {
 			return null;
 		}
+
+		session.removeAttribute(WebKeys.OPEN_ID_LOGIN);
+
+		User user = UserLocalServiceUtil.getUserById(userId);
 
 		String[] credentials = new String[3];
 
 		credentials[0] = String.valueOf(user.getUserId());
 		credentials[1] = user.getPassword();
-		credentials[2] = Boolean.FALSE.toString();
+		credentials[2] = Boolean.TRUE.toString();
 
 		return credentials;
 	}
 
-	protected User getUser(HttpServletRequest request, long companyId)
-		throws PortalException {
-
-		HttpSession session = request.getSession();
-
-		String emailAddress = (String)session.getAttribute(
-			WebKeys.FACEBOOK_USER_EMAIL_ADDRESS);
-
-		if (Validator.isNotNull(emailAddress)) {
-			session.removeAttribute(WebKeys.FACEBOOK_USER_EMAIL_ADDRESS);
-
-			return UserLocalServiceUtil.getUserByEmailAddress(
-				companyId, emailAddress);
-		}
-		else {
-			long facebookId = GetterUtil.getLong(
-				(String)session.getAttribute(
-					FacebookConnectWebKeys.FACEBOOK_USER_ID));
-
-			if (facebookId > 0) {
-				return UserLocalServiceUtil.getUserByFacebookId(
-					companyId, facebookId);
-			}
-		}
-
-		return null;
-	}
-
 	@Reference
-	protected void setFacebookConnect(FacebookConnect facebookConnect) {
-		_facebookConnect = facebookConnect;
+	protected void setOpenId(OpenId openId) {
+		_openId = openId;
 	}
 
-	private FacebookConnect _facebookConnect;
+	private OpenId _openId;
 
 }
