+++ modules/portal-security/portal-security-sso-opensso/src/com/liferay/portal/security/sso/opensso/auto/login/OpenSSOAutoLogin.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.opensso.security.auth;
+package com.liferay.portal.security.sso.opensso.auto.login;
 
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.log.Log;
