+++ modules/portal-security/portal-security-sso-token/src/com/liferay/portal/security/sso/token/auto/login/TokenAutoLogin.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.token.internal.security.auth;
+package com.liferay.portal.security.sso.token.auto.login;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
